use_cases/shared-files-through-system/shared-file-reader/requirements.txt
# This is where required Python libraries are listed
# pandas
# numpy
# scikit-learn
kelvin-app[data]>=6.0.0
pytest